Split off docker requirement tests from the big main testbench into its own
Move config & script into auxiliary_requirements
Move existing auxiliary to auxiliary_resource
