+++ src/test/resources/testbash/auxiliary_requirements/check_requirements.sh
