+++ src/test/resources/testbash/auxiliary_resource/check_bash_version.sh
