+++ src/test/resources/testbash/auxiliary_resource/code.sh
