+++ src/test/resources/testbash/auxiliary_resource/config_resource_test.vsh.yaml
