+++ src/test/resources/testbash/auxiliary_resource/config_resource_unsupported_protocol.vsh.yaml
