+++ src/test/resources/testbash/auxiliary_resource/resource1.txt
