+++ src/test/resources/testbash/auxiliary_resource/resource2.txt
