+++ src/test/resources/testbash/auxiliary_resource/resource3.txt
