+++ src/test/resources/testbash/auxiliary_resource/resource_folder/resource_L1_2.txt
