+++ src/test/resources/testbash/auxiliary_resource/resource_folder/resource_L2/resource_L2_1.txt
